+++ packages/odata-service-inquirer/src/prompts/datasources/sap-system/service-selection/questions.ts
@@ -232,7 +232,7 @@ export function getSystemServiceQuestion(
          */
         questions.push(
             ...withCondition(
-                [getValueHelpDownloadPrompt(connectValidator, promptNamespace, convertedMetadataRef)],
+                getValueHelpDownloadPrompt(connectValidator, promptNamespace, convertedMetadataRef),
                 (answers: { [serviceSelectionPromptName]?: ServiceAnswer }) =>
                     !!(connectValidator.serviceProvider instanceof AbapServiceProvider) &&
                     !!convertedMetadataRef.convertedMetadata &&