+++ packages/ui5-library-writer/templates/optional/typescript/eslint.config.mjs
@@ -1,0 +1,5 @@
+import fioriTools from '@sap-ux/eslint-plugin-fiori-tools';
+
+export default [
+    ...fioriTools.configs.recommended,
+];