+++ packages/ui-service-sub-generator/CHANGELOG.md
@@ -1,5 +1,78 @@
 # @sap-ux/ui-service-sub-generator
 
+## 0.1.183
+
+### Patch Changes
+
+- @sap-ux/axios-extension@1.25.11
+- @sap-ux/fiori-generator-shared@0.13.70
+- @sap-ux/inquirer-common@0.11.4
+- @sap-ux/system-access@0.6.50
+- @sap-ux/telemetry@0.6.71
+- @sap-ux/ui-service-inquirer@0.2.120
+
+## 0.1.182
+
+### Patch Changes
+
+- @sap-ux/ui-service-inquirer@0.2.119
+
+## 0.1.181
+
+### Patch Changes
+
+- @sap-ux/axios-extension@1.25.11
+- @sap-ux/fiori-generator-shared@0.13.69
+- @sap-ux/inquirer-common@0.11.3
+- @sap-ux/system-access@0.6.50
+- @sap-ux/telemetry@0.6.70
+- @sap-ux/ui-service-inquirer@0.2.118
+
+## 0.1.180
+
+### Patch Changes
+
+- Updated dependencies [2fc459c]
+    - @sap-ux/axios-extension@1.25.11
+    - @sap-ux/inquirer-common@0.11.2
+    - @sap-ux/btp-utils@1.1.8
+    - @sap-ux/telemetry@0.6.69
+    - @sap-ux/ui-service-inquirer@0.2.117
+    - @sap-ux/fiori-generator-shared@0.13.68
+    - @sap-ux/system-access@0.6.50
+
+## 0.1.179
+
+### Patch Changes
+
+- @sap-ux/ui-service-inquirer@0.2.116
+
+## 0.1.178
+
+### Patch Changes
+
+- Updated dependencies [1ad56d1]
+    - @sap-ux/inquirer-common@0.11.1
+    - @sap-ux/ui-service-inquirer@0.2.115
+
+## 0.1.177
+
+### Patch Changes
+
+- Updated dependencies [a5ecd7f]
+    - @sap-ux/inquirer-common@0.11.0
+    - @sap-ux/ui-service-inquirer@0.2.114
+    - @sap-ux/axios-extension@1.25.10
+    - @sap-ux/fiori-generator-shared@0.13.67
+    - @sap-ux/system-access@0.6.49
+    - @sap-ux/telemetry@0.6.68
+
+## 0.1.176
+
+### Patch Changes
+
+- @sap-ux/ui-service-inquirer@0.2.113
+
 ## 0.1.175
 
 ### Patch Changes