+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.6.63
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.4
+- @sap-ux/project-input-validator@0.6.55
+
+## 0.6.62
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.3
+- @sap-ux/project-input-validator@0.6.54
+
+## 0.6.61
+
+### Patch Changes
+
+- Updated dependencies [2fc459c]
+    - @sap-ux/inquirer-common@0.11.2
+    - @sap-ux/ui5-info@0.13.12
+    - @sap-ux/project-input-validator@0.6.53
+
+## 0.6.60
+
+### Patch Changes
+
+- Updated dependencies [1ad56d1]
+    - @sap-ux/inquirer-common@0.11.1
+
+## 0.6.59
+
+### Patch Changes
+
+- Updated dependencies [a5ecd7f]
+    - @sap-ux/inquirer-common@0.11.0
+    - @sap-ux/project-input-validator@0.6.52
+
 ## 0.6.58
 
 ### Patch Changes