+++ packages/ui5-test-writer/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @sap-ux/ui5-test-writer
 
+## 0.7.68
+
+### Patch Changes
+
+- Updated dependencies [346f09c]
+    - @sap-ux/project-access@1.35.3
+    - @sap-ux/ui5-application-writer@1.7.8
+
+## 0.7.67
+
+### Patch Changes
+
+- Updated dependencies [d2b772d]
+    - @sap-ux/project-access@1.35.2
+    - @sap-ux/ui5-application-writer@1.7.8
+
+## 0.7.66
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.1
+- @sap-ux/ui5-application-writer@1.7.8
+
+## 0.7.65
+
+### Patch Changes
+
+- Updated dependencies [2bffca7]
+    - @sap-ux/ui5-application-writer@1.7.7
+
+## 0.7.64
+
+### Patch Changes
+
+- Updated dependencies [a5ecd7f]
+    - @sap-ux/project-access@1.35.0
+    - @sap-ux/ui5-application-writer@1.7.6
+
 ## 0.7.63
 
 ### Patch Changes